Refactors Options to fix Rubocop ClassLength cop
